IO: Split STM32F4 specific code for serial interface.
