+++ src/drivers/boards/px4fmu-v4pro/board_config.h
@@ -60,6 +60,8 @@
 #define PX4IO_SERIAL_VECTOR     STM32_IRQ_USART6
 #define PX4IO_SERIAL_TX_DMAMAP  DMAMAP_USART6_TX_2
 #define PX4IO_SERIAL_RX_DMAMAP  DMAMAP_USART6_RX_2
+#define PX4IO_SERIAL_RCC_REG	STM32_RCC_APB2ENR
+#define PX4IO_SERIAL_RCC_EN	RCC_APB2ENR_USART6EN
 #define PX4IO_SERIAL_CLOCK      STM32_PCLK2_FREQUENCY
 #define PX4IO_SERIAL_BITRATE    1500000			/* 1.5Mbps -> max rate for IO */
