+++ src/drivers/px4io/CMakeLists.txt
@@ -39,6 +39,7 @@ px4_add_module(
 		px4io.cpp
 		px4io_uploader.cpp
 		px4io_serial.cpp
+		px4io_serial_f4.cpp
 	DEPENDS
 		platforms__common
 	)
